WENO4

Documentation for WENO4.

WENO4.WENO4

WENO4.binary_search

WENO4.interpolate_weno4

WENO4.weno4_q

WENO4.weno4_β

WENO4.WENO4 — Module

Interpolation routines for 4th order Weighted Essentially Non-Oscillatory (WENO) schemes.

WENO4.binary_search — Method

binary_search(a::AbstractVector{<:Number}, x)

Return the index of the last value in a less than or equal to x. a is assumed to be sorted.

WENO4.interpolate_weno4 — Method

function interpolate_weno4(
    xs::AbstractVector{<:Number},
    xp::AbstractVector{<:Number},
    fp::AbstractVector{<:Number};
    extrapolate=false
)

Performs 1D interpolation using the fourth-order Weighted Essentially Non-Oscillatory (WENO) scheme from Janett et al (2019). Based on https://github.com/Goobley/Weno4Interpolation by Chris Osborne.

Arguments

xs::AbstractVector{<:Number}: The x-coordinates at which to evaluate the interpolated values

xp::AbstractVector{<:Number}: The x-coordinates of the data points. Must have at least four elements, be increasing, and have no NaNs.

fp::AbstractVector{<:Number}: The y-coordinates of the data points. Must have same length as xp.

Keywords

extrapolate::Bool: whether or not to use extrapolation. If true, will perform quadratic extrapolation from the edge points. When false, it will not throw an error if xs is outside xp, but return the closest fp value.

WENO4.weno4_q — Method

Compute the q2 and q3 Lagrange interpolation factors for 4th order WENO interpolation. Arguments are the value to interpolate plus the x and y values for the 4-point stencil.

WENO4.weno4_β — Method

Compute the smoothness indicators β2 and β3 for 4th order WENO interpolation. Arguments are the x and y values for the 4-point stencil.